+++ R/rowwise.r
@@ -33,7 +33,7 @@ print.rowwise_df <- function(x, ..., n = NULL, width = NULL) {
 
 #' @export
 ungroup.rowwise_df <- function(x) {
-  class(x) <- "data.frame"
+  class(x) <- c( "tbl_df", "data.frame")
   x
 }
 #' @export
@@ -101,5 +101,3 @@ do_.rowwise_df <- function(.data, ..., .dots) {
     label_output_list(NULL, out, groups(.data))
   }
 }
-
-